main,refactor: split dependency.h into dependency.h and dependency_p.h
