+++ main/dependency.h
@@ -12,11 +12,17 @@
 #ifndef CTAGS_MAIN_DEPENDENCY_H
 #define CTAGS_MAIN_DEPENDENCY_H
 
-#include "general.h"
+/*
+*   INCLUDE FILES
+*/
+#include "general.h"  /* must always come first */
 
 #include "types.h"
-#include "kind.h"
 
+
+/*
+*   DATA DECLARATIONS
+*/
 typedef enum eDepType {
 	DEPTYPE_KIND_OWNER,
 	DEPTYPE_SUBPARSER,
@@ -35,29 +41,5 @@ struct sSlaveParser {
 	void *data;
 	slaveParser *next;
 };
-
-struct slaveControlBlock;	/* Opaque data type for parse.c */
-
-
-extern void linkDependencyAtInitializeParsing (depType dtype,
-						   parserDefinition *const master,
-						   struct slaveControlBlock *masterSCB,
-						   struct kindControlBlock *masterKCB,
-						   parserDefinition *const slave,
-						   struct kindControlBlock *slaveKCB,
-						   void *data);
-
-extern struct slaveControlBlock *allocSlaveControlBlock (parserDefinition *parser);
-extern void freeSlaveControlBlock (struct slaveControlBlock *cb);
-extern void initializeDependencies (parserDefinition *parser,
-									struct slaveControlBlock *cb);
-extern void finalizeDependencies (parserDefinition *parser,
-								  struct slaveControlBlock *cb);
-
-extern slaveParser *getFirstSlaveParser(struct slaveControlBlock *controlBlock);
-extern slaveParser *getNextSlaveParser(slaveParser *last);
-#define foreachSlaveParser(VAR)			\
-	VAR = NULL;								\
-	while ((VAR = getNextSlaveParser (VAR)) != NULL)
 
 #endif	/* CTAGS_MAIN_DEPENDENCY_H */
