+++ win32/ctags_vs2013.vcxproj
@@ -253,6 +253,7 @@
     <ClInclude Include="..\main\ctags.h" />
     <ClInclude Include="..\main\debug.h" />
     <ClInclude Include="..\main\dependency.h" />
+    <ClInclude Include="..\main\dependency_p.h" />
     <ClInclude Include="..\main\e_msoft.h" />
     <ClInclude Include="..\main\entry.h" />
     <ClInclude Include="..\main\entry_p.h" />
